Rewrite Dockerfile: Debian/glibc base, drop sqlean rebuild and wasm tooling
