+++ core/testlib/Cargo.toml
@@ -18,3 +18,6 @@ tracing = { workspace = true }
 tempfile = { workspace = true }
 serial_test = { workspace = true }
 blst = "=0.3.16"
+
+[build-dependencies]
+glob = { workspace = true }
